rating is bigDecimal

+++ driver-service/src/main/kotlin/com/example/driver_service/model/dto/DriverResponseDto.kt
@@ -1,6 +1,7 @@
 package com.example.driver_service.model.dto
 
 import com.example.driver_service.model.enums.Gender
+import java.math.BigDecimal
 import java.util.UUID
 
 data class DriverResponseDto(
@@ -8,7 +9,7 @@ data class DriverResponseDto(
     val name: String,
     val email: String,
     val phoneNumber: String,
-    val rating: Float?,
+    val rating: BigDecimal?,
     val gender: Gender,
     val carId: UUID?,
 )
